+++ recipes/config_https.rb
@@ -10,9 +10,17 @@ ruby_block 'configure https' do
       aquarius_data_bag = data_bag_item(node['aq_config']['databag'], node['aq_config']['database_databag'])
       cert_pass = aquarius_data_bag['cert_password']
       cert_name = aquarius_data_bag['cert_name']
-      `\"C:\\Program Files\\Amazon\\AWSCLI\\aws\" s3 cp \"s3://devops-owi-configuration-management/application/aquarius/configuration/asg_certs/#{cert_name}\" \"C:\\#{cert_name}\"`
-      configout = `\"C:\\Program Files\\Common Files\\Aquatic Informatics\\AQUARIUS\\HttpsConfigurationTool.exe\" -c\"C:\\#{cert_name}\" -p#{cert_pass}`
-      if configout.include? 'ERROR: Unable to configure HTTPS.'
+      tier = node['aq_config']['tier']
+      if File.file?("C:\\Program Files\\Common Files\\Aquatic Informatics\\AQUARIUS\\HttpsConfigurationTool.exe")
+        httpstool = "C:\\Program Files\\Common Files\\Aquatic Informatics\\AQUARIUS\\HttpsConfigurationTool.exe"
+      elsif File.file?("C:\Program Files\\Aquatic Informatics\\AQUARIUS Server\\bin\\HttpsConfigurationTool.exe")
+        httpstool = "C:\Program Files\\Aquatic Informatics\\AQUARIUS Server\\bin\\HttpsConfigurationTool.exe"
+      else
+        Chef::Log.error('Unable to find HttpsConfigurationTool.exe!!!')
+      end
+      `\"C:\\Program Files\\Amazon\\AWSCLI\\aws\" s3 cp \"s3://devops-owi-configuration-management/application/aquarius/configuration/#{tier}/asg_certs/#{cert_name}\" \"C:\\#{cert_name}\"`
+      configout = `\"#{httpstool}\" -c\"C:\\#{cert_name}\" -p#{cert_pass}`
+      unless configout.to_s.empty?
         Chef::Log.error('Unable to configure HTTPS')
       end
     end